Add dsa_pub.pem file so things build.
The file is actually useless -- it would appear be used to verify
updates from a now non-existent source -- but until things get fixed
it is useful to have it.

This dsa_pub.pem file comes from a binary copy of Fairmount. Possibly
this is a bad idea and a complete dummy should be here instead.