Fix and clean up library search code and user prompting.
Before, it was assumed the needed library was installed in the VLC.app
bundle. Now, however, it needs to be downloaded on its own from
videolan.org. I've rigged it so that the dialog now gives the user the
opportunity to get the .pkg file needed directly and retry.

We now also search a sequence of possible library locations.

I've also removed a lot of code for prompting the user for where the
VLC.app bundle is installed, for searching the VLC.app bundle, etc.,
since none of that is now needed.