Changed tabs to whitepace, killed trailing whitepace, etc.
The repair was done by running the script Tools/fixtabs.sh (added to
the repository). It converts tabs to spaces with four column tab stops
(which was the apparent intent), removes trailing whitepace, fixes a
file with \r\n DOS-style line endings, and cuts the length of the
separator comments down below 80 columns.

If you have a divergent codebase, you can regularize it with the same
script, and then diffs will be minimized.

I did this with great reluctance, having had such changes blow up in
my face in the past, but there were too many places where spaces were
mixed badly with tabs, and it was generally too hard to look at the
code in emacs or on github or in the shell.

Hopefully since I seem to be the sole person hacking on this at the
moment it won't be too disruptive -- I decided to do it now before
anyone else works on it rather than later when it would be
impossible.